Integrate the transactional backend
Each bunch is now submitted as a whole to opkg-trans. This is an
intermediate step in the direction to the real updater-ng, integrating
the backend part.

The journal recovery is somehow halfway (we abort if it doesn't work the
first time around). Also, it is very verbose in logs for now (we want to
watch what it does).

As we have both opkg and opkg-trans, there's a change we can get into
inconsistent state if opkg installs removes something in between
interrupted opkg-trans run and its journal recovery. This shouldn't
happen in practice much (as the interrupt would be power off and
recovery happens at boot) and it is a temporary problem anyway.