Updater CLI arguments: Add error message operation
Instead of using COT_CRASH for error message, we are now using new
operation. This way we can print multiple error messages and at the end
we don't have to use any dynamically allocated memory, because we are
not merging constant strings with arguments any more, we print them as
separate messages.