Add update hooks to pkgupdate
We want to invoke snapper every time we are going to perform
update. These hooks are introduced to implement this requirement in some
general way.

We implement two hooks for now. Preupdate hook is executed before
updater does any changes and postupdate hook is executed after updater
done all planned changes.

Tests for this are added as part of system plan test.