backend: Install dead symbolic links
If there's a dead symbolic link in between the files to be installed
(because the target has already been installed), then we want to use
lstat to determine the existence of the link, instead of stat (which
would say it doesn't exist). That way the symbolic link is installed as
well.