Add sat as lua extension for picosat
This new extension allows adding clauses in form of batches. Single
batch then can or don't have to be added to picosat. This gives us
ability to drop some set of clauses when we encounter missing package
and we should ignore it.

Batches can refer to other batches. If batch is committed so are all
referred batches.