luacheck: define globals in files it self
We define all globals options for luacheck in file it self rather then
in global configuration file. Not only that this is cleaner with
respect to adding new lua scripts or renaming old ones, but also this
list of globals is in file it self and don't have to be edited
separately.