Build fix: make make less confused about filter-out
It seems the filter-out rules were expanded at different times,
producing different output than when creating the prerequisites. So it
didn't filter them out.

Now we leave the first one out by $< ‒ which must be the same. We leave
the second one in there. But that one is used by the script ‒ we remove
it from listing it explicitly. A bit of a hack, but it seems to work
better.