Add script for delayed reboot hook and run updater.sh in init
Script added is going to be used as reboot_required hook. It creates
message requesting reboot.

Second change is that when we recover journal in init, we will run
updater.sh. This solves situation with lost replan request. If update
was finished and there are no more changes, running updater.sh will have
no effect.