Provide a priority to standalone packages
Allow the Package to have a priority option and use it when comparing
stand-alone packages with other packages to pick the right one.

This'll allow us to have a testing package with a version higher than
the one in repository and have the repository outgrow the version in
future (or not, depending on the priority).