When running command run it in separate process group
This is partial fix to problem of shell masking kill commands to its
children. They wouldn't die and because we wait for process to close
streams with us we hand even if we SIGKILL it. Another fix should be
that we shouldn't be waiting for streams to close and instead close them
on our own if we do SIGKILL, but that will be more complicated.