Add no-replan option to pkgupdate
When we are not using pkgupdate from packages (for example when we are
running it in different root) then there is no reason to do replan. It
might even cause problems when we are installing clean to clean root.
This new option allows can be specified to deny replanning and that way
ensure that everything is installed at once.