Multiple packages with Provides should be expected to be used
When we have two packages providing same package and both should be
installed, than we should handle them the same way. Until now we have
chosen first candidate that was selected and that was the one that we
used. But that is problematic if we want to install both packages.

This commit ensures that when we request a package that is provided by
more than one other package than we pass modifier to both of those
packages instead of random single one.