Drop no longer relevant documentation
Drop migration documentation

Drop ideas documentation as we are far from that stage now
All ideas described here are implemented or not going to be. This file
now adds nothing to documentation.

Drop battleplan as that file is no longer relevant.

Remove morpher documentation as it's deprecated and will be removed.