Remove need_approval file if plan is clean
There is possible situation when we generated approval plan and then
disabled approvals which leads to update to pass. When we reactivate
approvals it is reported as new plan to approve, but there is no such
change planned, it's just a ghost of old plan. This ensures that we
remove plan when we have no new plan.