Split all functionality off entry.lua and do some cleanups
In next updater release will localrepo and opkg wrapper be in separate
packages and because we now have conf.d directory we can easily add any
additional configuration code trough it and package it in separate
package. This makes much more clean entry script.

Major changes with this is that auto.lua and user.lua are no longer
loaded and we have to migrate them to /etc/updater/conf.d. So following
move will be done:
/etc/updater/user.lua -> /etc/updater/conf.d/user.lua
/etc/updater/auto.lua -> /etc/updater/conf.d/opkg-auto.lua