Fix invalid order of candidates in some cases
There was a bug in compare function for candidates sorting. Sorting
algorithm is quicksort and requires that compare returns true when first
is less then the second argument. This implies that we should return
false as last resort (for comparable candidates). We were returning true
which resulted to invalid order in some cases.