Don't validate uri options when they are inserted to context
It makes little to no sense to validate extra arguments when they are
pushed to sub-script context as if they would be invalid the target
script would already failed. This is just useless check and would add
complexity as it makes table from any argument passed to it so this way
it discards uri.system_cas and uri.no_crl and creates bugged system.