Run postupdate hooks as part of cleanup when --replan is used
There is situation when we are replanning when we end up with no new
changes to be done and this way we exit early (going to cleanup). That
is problematic as that means that we won't run postupdate hooks in such
situation. This fixes it. When we go to cleanup and we have --replan we
also run postupdate hooks.