Fixing problems with localrepo
There is multiple problems with localrepo.
Command in file localrepo.lua were generated on single line not ending
with new line character.
Empty repository wasn't added to localrepo.lua as there was no
Packages.gz file for it. Although that might be fine as we expect that
it doesn't have to exist but we generate localrepo.lua only if we are
creating or removing repositories. That caused situation when latest
repository was never added to localrepo.lua as it was empty at the
moment of localrepo.lua generation.
Because of that when we are trying to fix localrepo repository we also
regenerate automatically localrepo.lua just to have a way to request
that.