WIP: Use cleanup for postupdate hooks
We had previously problem when updater died and it didn't executed
postupdate hooks. This should solve it. At least for cases when updater
fails in some sensible way.

Note: When we run replan then we should unregister postupdate hook but
problem is that at the moment replan is triggered from libupdater and
from lua. We should move that to pkgupdate instead a probably to C.