Fix migration to supervisor
When updater is updated then foris can approve another execution but
because it's not yet updated it runs updater.sh. But because we don't
have that installed yet we run pkgupdate directly instead.

This has side effect that migration update won't report changes to
notification system. But that should be fine.