Fallback to pkgupdate when -b is passed as argument
Durring migration a -b -c is passed to updater.sh as arguments. This
causes pkgupdate to fail. We should filter those out and if -b is
specified we should run updater-supervisor instead. But there is a case
when we don't have it installed yet and because of that we can allso
fallback to pkgupdate directly.