Allow usage of os module in local security level
It already has access to io and file. There is also no sense to limit
access to os as as discussed in commit adding io and file if you get
local file on router then you already have to have some kind of access
to router it self. And if Local access is miss used then there is no
help either way.