Do not build entry.lua to pkgupdate
entry.lua is more or less configuration and in reality we don't need it
for any migration. It no longer has anything relevant inside and
everything is in directory /etc/updater/conf.d anyway. Note about
migration makes no sense in current setup as pkgupdate is in the same
package as this configuration file.