opkg-wrapper: allow ignoring custom feeds
This adds ignore list for opkg-wrapper. It is implemented using Lua
pattern matching function. If one of provided patters is matching given
line in customfeeds.conf then feed from such line is ignored. Source for
used patterns are files in a directory '/etc/updater/opkg-ignore'.
Every line is a patterns with exception of those starting with '#',
those are intended for commenting.

There is a situation where custom feed is added and to allow it to be
easily used it is added not only to opkg but also to updater. In such
case such feed is pulled twice because of opkg-wrapper. This would not
be huge problem unless there are some additional settings and both
downloads are not in fact same. This can decrease security by going
around of more secure updater's configuration and less secure URL from
opkg-wrapper.

Unfortunately we can't do this automatically because there is no access
to already known repositories. And even if such access would existed we
couldn't be sure that repository we are adding is not going to be added
later then opkg-wrapper is executed. Because of that much more better
solution is to have some sort of ignore list for opkg-wrapper.