Rename opkg-trans to pkgtransaction
This is huge change but it makes sense. Original name is now pretty
confusing and it should be clear that this tool is part of updater-ng.
This new name should signal that.
This new name also contains whole word "transaction" so it should
possibly be more clearer what it really does. This also adds comment
that this tool is dangerous to play with.