supervisor: update to allow autorun configuration
This adds possibility to completely configure updater (autorun) trough
supervisor. Advantage of this is that supervisor is the only place that
defines uci configuration for updater (at least if we are talking about
autorun) and because of that same API can be used on top of generally
different configuration. This is handy because of back compatibility
with Turris OS 3.x and introduced cleanup in Turris OS 4.x.