logging: drop unneeded exit code
I probably had some plan with it but I no longer remember what it was
suppose to been. Instead I just drop it here. It can be reverted later
if needed.

This also correctly sets read field for cookie structure. This has no
effect because we are not reading from it but just to be precise with
our code we set it to know value NULL which means that read is not
available.