Add multiwrite
This is simple wrapper on top of multiple opened files. Idea is that you
can open additional files and all subsequent writes are written to all
of those files.
The intended use is for collecting CA and CRL certificates to single
file. In that implementation we can have multiple nodes reusing same
base list of CAs and CRLs. To allows deallocation of deployed
certificates we create files that can be read later when ever we detect
such reuse. This simple code makes this operation as easy as writing to
single file descriptor.