transaction: instead of data use files on fs
This replaces problematic data pass in Lua with files saved on disk.
Original problem is that Lua in some cases like closures duplicates
environment and that means also all data. If this happens few times
(from measurements it seems like two or three times in code segment we
are dropping in this) then we can have few times bigger memory
requirements.

This also temporally disables test test-sys-pkgupdate-plan-unapproved as
it does not expect additional files. This is going to be solved in
following commit.