Intefate new URI implementation to Lua code
This changes a lot of behavior. The most significant from users point is
that in configuration language functions no longer return handlers for
configuration they created. This is to simplify implementation. The
correct approach is to use package or repository name instead of
handler. Handler usage was less powerful and because of sandboxing also
unusable across multiple scripts.