doc: drop plan of package pre_* post_* function hooks
This was never implemented and it somewhat questionable. It might be
interesting in future to reintroduce it back and instead of having it as
a lua function to have it as shell script (run by something like
system()).

Why this is dropped is because it is questionable on how to implement
it. It requires to run function from script that is no longer available
potentially. It has to be run in transaction which has nothing to do
with original configuration scripts. This was just stupid design idea
and we should not implement it.