uri: add support for gzipped content signature verification
With indexes we have a problem that they can be gzipped but signature is
generated against non-gzipped file. This means that checking downloaded
gzipped content against signature it won't match. The only approach I
have come up with is to try first verify if content is signed and if
that fails then we check if there is an gzip magic at the beginning and
in such case we check gunzip file and try again. It is not ideal but in
general it is good workaround to solve this nasty hack OpenWRT is doing.