backend: remove all files early
We encountered a problem with late files removal. Those removals were
clashing with postinst scripts. Postinst script expect to be run on
fully updated system and having leftover files from older version was
not expected. This had effect like removal of files that were generated
by postinst or postrm scripts or uninstalled plugins being loaded to
applications on service restart by postinst.

This just simply reuses existing facility to remove files early during
files merge. Instead of using it to remove files that collide with
directories we use it to remove all abundant files. It is not an ideal
solution but we are unable to drop previous approach because of backward
compatibility with previous journals.