openssl_purity: simplify makefile
This just refactors original openssl_purity.mk.

Also note that this might no longer be required. From OpenSSL 1.1.0 it
seems that problems with valgrind were fixed and this might meant that
we can drop this. We don't want to do that now because OpenSSL 1.0.0
only systems might still exists but we can do it in future.