uri: make threads conditional
We do not need error variables to be thread_local but it would be good
to set them so. The problem is that some libc implementations with C11
support do not provide this header. We can check for that and for now
just ignore given option. This means that we provide it in case of
existing threads.h and ignore it in other case.