tests: drop events tests
These tests are regularly failing because they are implemented in a
strict time sensitive way. The reason why we can drop them is because we
are pretty much no longer using it. We are planning to drop it as soon
as we replace utilities calls with something different. That is the only
feature we are still using from events and that specific feature wasn't
even tested in test we are dropping.