lib: drop possibility to embed busybox
We do not need this. This was originally introduced to ensure that
updater on its own can run and update system that is pretty much broken.
The problem is that on such system we fail to run as well. Busybox has
subset of our dependencies and when we are able to run then busybox
binary is able to do that as well. We can also update busybox as part of
our dependencies same as for example libc and there is no gain in not
doing that.
This also had a problem that we were fixed on /tmp usage. Any package
could potentially cause remount of /tmp or could remove content of it
and updater would no longer work. That means that this was making
updater potentially more unstable compared to using system busybox.