pkgupdate: update alternatives hook fix to use ROOT_DIR
This fixes problem where alternatives were not correctly created in
system run in non-root mode. This was because as source the global path
was used but in places like build system this path is invalid. We have
those files but relative to ROOT_DIR path. This fixes that.

This also just fixes small shell syntax problem where paths were no
quoted.