pkgupdate: create links to alternatives in correct path
The previous fix correctly changed source of alternatives but it was
still trying to create link in invalid location (ignoring configured
root). This adds this path in appropriate location.

Note that TRG is an absolute path so there does not have to be any slash
between ROOT_DIR and TRG.