lib/autoload/postprocess: do not consider provides of same name
This warns and ignores provides for packages of same name. The reason is
that otherwise we would add copy of package candidates to list of
candidates making it appear twice. The result together with dependency
resolution is that this candidate would conflict with itself and
couldn't be installed at all with updater.