Make virtual packages just fulfilled automatically
This changes behavior of virtual for packages. With this implementation
virtual packages no longer produce error when there is existing package.
It instead just ignores those packages.

The change from original behavior on top of just not throwing error is
that provides for virtual package are not considered at all.

This technically standardizes behavior to state where virtual package is
just considered as fulfilled without any candidate and no candidate is
installed for it. You can mark any package as virtual to have it
considered by updater as installed. This skips all candidates (including
their dependencies) and only dependencies introduced from updater's
configuration are required.

This adds new feature symbol reported to user: no_error_virtual
The reason is to differenciate between updater versions before and after
this change to virtual behavior.