lib: allow requests to be conditioned
This adds new extra "condition" to Install and Uninstall requests. Those
are intended to solve problem of cyclic dependencies if we have package
that extends some other but such dependency has to be patched to that
package and commonly extending package should rather depend on package
it extends not other way around.

Doing it this way allows us to do magic such as that user configures
updater to track specific languages and exact packages are installed
depending on other requested packages.

This behavior can't be solved any other way cleanly because extending
package depends on package itself. If we do it so that if given package
is installed that we request extending package then we end up with
package lock. Package is never removed even if it is no longer required
unless explicitly uninstalled.