lib/archive: implement decompress function
This is used to decompress repository indexes. This implementation is
pure C and uses libarchive.

This allows us to decompress any format that is supported by libarchive
but we currenctly use it on GZIP only. We can remove check for magic
bytes later on and just decompress any other format but that is not
required now as opkg is not supporting that.