src/lib/archive: add unpack_package and error handling
This now implements package unpack using libarchive instead of calling
tar. This was done primarily to solve memory requirements needed to
unpack big packages to memory. Instead of that we now do it in stream
and that is way less memory consuming.

Design is based around FILE descriptors and in general around data
streams.

unpack_package is immediately used in backend.lua to replace original
implementation.