update_alternatives: fix problem with busybox
Busybox is not using alternatives but provides same links as some
alternatives. To fix any removed links we first recreate all busybox
links and then overwrite any of them.

This also removes unnecessary sed that is no longer required to reorder.

The second purpose was to remove space at the beginning but that is not
required with sort -n as that ignores spaces to do numeric sort and
space is removed by read later on.