lib: add signature check implementation using OpenSSL
This replaces call to usign that was suboptimal. Every message,
signature and public key had to be dumped to FS somehere to call usign
on it. Usign supports one of files to be piped in but it works the way
that it first reads it all and then evaluates and only one part of three
can be provided this way.

This implementation operates directly on buffer. The advantage of that
is that it does not necessary have to be in memory if we use mmap. The
advantage is also that there is no need to dump anything to FS.
Everything is directly provided in program.