tests/c/archive: fix comparison test
Comparing against files tracked in git is not ideal as git is benevolent
about access rights. It tracks execution but uses umask to mask group
and other permissions. In the end this means that we can have different
access rights on different system.

Replacement is now rather hacky but it rather compares tar output with
our own. This should ensure that we behave in somewhat consistent way
without relying on git files representation.