lib/download: remove download retries
Download retries are flawed because if some data is downloaded and later
download is started again the new data is appended to old data. This
means that result is invalid even if new download was successful.

One option would be to truncate output for retry but output can be not
only file but effectively any buffer. I am not aware of easy way to
truncate FILE. We can truncate real file but it seems that there is no
way to truncate memory objects such as created using fopencookie.

It is also questionable if retries were even usable for anything. Curl
on its own tries to reestablish connection and letting it on Curl is in
general even cleaner than doing some hacks.