lib: fix invalid error for uri_content
It is not possible to universally call :path() on URIs as not all of
them are of file scheme. Instead :uri() should have been called here to
get URI representation.

This also not essentially downloading URI. It just receives it no matter
what kind of URI it is so message about download is miss leading
(although in most cases valid).