lib: replace libb64 with openssl
We have problem with libb64 on ARM (it effectively does not work).
Updater is already linked against OpenSSL so why not use it for that as
well. The API for it is not exactly clean so it is wrapped in its own
file but otherwise it is clean solution as it removes dependency and
reuses existing one.