Remove immediate reboot
This removes immediate reboot possibility from updater.

Original reasoning was to reboot router because update of specific
package broke system so much it can't continue. We never used it. It is
also questionable how such package could break system in a way where
reboot helps. That would have to be kernel module but having some core
systems in module that we would have to do reboot and were unable to
finish update seems like impossible. To finish update we in general need
just filesystem and shell. Existing solution would not work with
theoretical breakage of any of those anyway so it is pretty much useless
feature. Anything this can solve can be also solved by reboot at the end
of update.