lib/lua/updater: fix broken queue messages about upgrade and downgrade
The code used for checking current system status was invalid and in
result whole if condition was being skipped silently. This now enabled
"upgrade", "downgrade" and "reinstall" messages instead of "install".

This also adds current version to upgrades and downgrades of packages.
This is info that is beneficial to know in most cases so it makes no
sense to not show it when we have simple access to it.