pkgupdate: fix missing ROOT_DIR in postupdate hook on replan
The previous expectation that ROOT_DIR is already set is true only if
current execution is reexecution (subsequent execution after replan). In
case of reexecution the ROOT_DIR set is skipped so we have to add it for
postinst as well.

That applies for both postupdate as well as reboot required hook.