tests/c/subprocess: fix potential test fail
It is possible that if we try to immediately kill process that it won't
make it to exec and that way kill might be caught by check library and
reported as failure. By waiting at least short amount we pretty much
make sure that exec happens and that we are killing sleep process rather
our fork.