lib/lua/transaction: fix postinst configure always reporting upgrade
The issue here was that control fields in status were updated in
pkg_move function but same and thus invalid check was used in
pkg_scripts later on.
This adds new table with all upgraded packages. This way we check
original state before it is removed from status and pass this info to
subsequent pkg_scripts.