Add FilesSignature support
We need this to cover case when files change in package without anything
other changing. The issue this tries to cover is conflicts between
packages because two packages provide same file but in reality that is
only because local version is older than the one in repository. This
forces reinstall and thus resolved such collision.