Replace base64 implementation with base64c library
In reality base64c library is based on this code so it is pretty much
same code just moved to separate library. The reason for the move is
that we want to use it in other projects as well and having it as
dedicated library should give us less code to maintain and fix.